requirement.txt
certifi==2018.11.29
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.3
django-markdownx==2.0.28
django-nested-admin==3.1.2
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
drf-yasg==1.11.1
flex==6.13.2
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
Markdown==3.0.1
MarkupSafe==1.1.0
openapi-codec==1.3.2
Pillow==5.3.0
PyJWT==1.7.1
python-monkey-business==1.0.0
pytz==2018.7
PyYAML==3.13
requests==2.21.0
rfc3987==1.3.8
ruamel.yaml==0.15.81
simplejson==3.16.0
six==1.11.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.1
uritemplate==3.0.0
urllib3==1.24.1
validate-email==1.3
